Remove the syntactic sugar analogy
The intent with that analogy was to describe how *part* of `addTeardown()`
is syntactic sugar over immediately checking `subscriber.signal.aborted`,
and *otherwise* setting the given cleanup function to the signal's abort
handler, and also invoking it during `complete()` and `error()`, but that
latter part was not clear at all and was contradictory as @annevk pointed
out. So this commit removes the example, in favor of the earlier descriptive
text.